requirements.txt
Django>=4.2
psycopg2-binary==2.9.9
python-dotenv==1.1.0
asgiref==3.8.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
firebase-admin
Pillow==11.2.1